Extract ColorPropertySet to separate files.
