+++ robot_array_rviz_plugins/src/SingleRobotStateArrayDisplay.h
@@ -15,43 +15,14 @@ namespace rviz
 class StringProperty;
 class BoolProperty;
 class IntProperty;
-class FloatProperty;
 class RosTopicProperty;
 class EnumProperty;
-class ColorProperty;
 } // namespace rviz
 
 namespace RobotArrayRvizPlugins
 {
 class SingleRobotStateArrayDisplay;
-
-class ColorPropertySet : public rviz::Property
-{
-  Q_OBJECT
-
-public:
-  ColorPropertySet(const std::string & name, SingleRobotStateArrayDisplay * display, rviz::Property * parent_property);
-
-  void setName(const std::string & name);
-
-  void setHidden(bool hidden);
-
-  bool getOriginal();
-
-  Ogre::ColourValue getColor();
-
-private Q_SLOTS:
-  void changedOriginal();
-
-protected:
-  SingleRobotStateArrayDisplay * display_;
-
-public:
-  rviz::Property * label_property_;
-  rviz::BoolProperty * original_property_;
-  rviz::ColorProperty * color_property_;
-  rviz::FloatProperty * alpha_property_;
-};
+class ColorPropertySet;
 
 class SingleRobotStateArrayDisplay : public rviz::Display
 {
